+++ lib/src/constants.dart
@@ -9,11 +9,10 @@ extension Opponent on Colour {
 }
 
 typedef Hand = List<int>;
-typedef FenBuilder = String Function();
 
 // TODO: there's a lot of stuff here now, maybe refactor and not call this constants?
 class Bishop {
-  static const version = '1.1.2';
+  static const version = '1.2.2';
 
   static const Colour white = 0;
   static const Colour black = 1;